added: awm-note, updated: compton-opacity

+++ zsh/.zshrc
@@ -188,16 +188,20 @@ man() {
 compton-opacity() {
     compton_config=~/.config/compton/compton.conf
     if [[ -e  $compton_config ]]; then
-        # change the opacity value in the 4th line
-        sed -i "4s/[0-9][0-9]/$1/" $compton_config
-        echo -e "$COLOR_HL1::$COLOR_TITLE urxvt transparency in compton.conf has been set to:$COLOR_HL1 $1 $COLOR_DEFAULT"
-        echo -e "$COLOR_HL1::$COLOR_TITLE Restart compton? (y = yes) $COLOR_DEFAULT"
-        read compton_restart
-        if [[ $compton_restart == "y" ]] || [[ $compton_restart == "Y" ]]; then
-            pkill compton &&
-            sleep 1s &&
-            compton -b --config $compton_config &&
-            echo "Compton restart: done!"
+        if [[ -z $1 ]]; then
+            echo "Error! Opacity value cannot be empty. You need to set a number."
+        else
+            # change the opacity value in the 4th line
+            sed -i "4s/[0-9][0-9]/$1/" $compton_config
+            echo -e "$COLOR_HL1::$COLOR_TITLE urxvt transparency in compton.conf has been set to:$COLOR_HL1 $1 $COLOR_DEFAULT"
+            echo -e "$COLOR_HL1::$COLOR_TITLE Restart compton? (y = yes) $COLOR_DEFAULT"
+            read compton_restart
+            if [[ $compton_restart == "y" ]] || [[ $compton_restart == "Y" ]]; then
+                pkill compton &&
+                sleep 1s &&
+                compton -b --config $compton_config &&
+                echo "Compton restart: done!"
+            fi
         fi
     else
         echo -e "$COLOR_HL1::$COLOR_TITLE Error! $compton_config doesn't exist. $COLOR_DEFAULT"
@@ -762,6 +766,58 @@ mailthis() {
     else
         echo "$message" | mutt -F ~/.mutt/account.1.muttrc -s "$subject" "$to"
     fi
+}
+
+# note file for using with awesome WM widget
+# widget format should be the same as in awesome rc.lua
+awm-note() {
+
+    NOTE_FILE=$HOME/Documents/awm-note.txt
+
+    if [[ ! -e $NOTE_FILE ]]; then
+        touch $NOTE_FILE
+        echo "Note file created at $NOTE_FILE"
+    fi
+
+    case "$1" in
+
+        # Listing notes
+        -l) cat $NOTE_FILE
+            ;;
+        # Delete line from notes
+        -r) sed -i ${2}d $NOTE_FILE
+            echo "removed line $2 from $NOTE_FILE"
+            ;;
+        # Delete ALL notes (delete the file)
+        -R)
+            echo "Delete ALL notes? [yY/nN]"
+            read answer
+            if [[ $answer == "y" ]] || [[ $answer == "Y" ]]; then
+                #echo -n > $NOTE_FILE
+                rm $NOTE_FILE
+                echo 'note_widget:set_text("")' | awesome-client
+                echo "All notes was deleted."
+            else
+                echo "Cancelled. Nothing changed."
+            fi
+            ;;
+        # Help
+        ""|-h) echo "Usage: awm-note [OPTIONS]"
+            echo -e "\nTo add a note, use:"
+            echo -e "\t awm-note \"your note in qoutes\""
+            echo -e "\nOptions:"
+            echo -e "-l\t\t list notes"
+            echo -e "-r [number]\t delete note (by line number)"
+            echo -e "-R \t\t delete ALL notes (delete the file)"
+            echo -e "-h\t\t this help"
+            ;;
+        # Add note
+        *)  echo "$1" >> $NOTE_FILE
+            echo "added '$1' to $NOTE_FILE"
+            echo 'note_widget:set_markup(" <span font_weight=\"bold\" color=\"red\">notes</span>")' | awesome-client
+            ;;
+    esac
+
 }
 
 # vless (use vim as a pager)
